FINALLY GOT AUTHENTICATION AND PERMISSIONS WORKING

+++ apiary/beekeeping_app/views.py
@@ -3,10 +3,27 @@ from django.http import HttpResponse
 from .models import Apiary
 from .models import Keeper
 from .models import Hive
-from .models import Event
+#from .models import Event
 
 from .forms import HiveForm
 from .forms import ApiaryForm
+
+#authentication
+from django.contrib.auth import authenticate, login, logout
+from .decorators import unauthenticated_user, allowed_users
+from django.contrib import messages
+from django.contrib.auth.models import Group
+from .forms import CreateUserForm
+from .forms import BeekeeperForm
+#this is for class-based views:
+from django.contrib.auth.mixins import LoginRequiredMixin
+#I'm using the login_required tag to achieve the same thing
+
+#permissions
+from django.contrib.auth.decorators import login_required, permission_required
+from django.core.exceptions import PermissionDenied
+from .decorators import allowed_users
+
 from django.shortcuts import redirect
 
 #for resizing the image as it's recieved
@@ -14,11 +31,11 @@ from PIL import Image
 from io import BytesIO
 
 #calendars:
-from datetime import datetime
-from django.views import generic
-from django.utils.safestring import mark_safe
-from .utils import Calendar
-from .models import Event
+# from datetime import datetime
+# from django.views import generic
+# from django.utils.safestring import mark_safe
+# from .utils import Calendar
+# from .models import Event
 
 
 # Create your views here.
@@ -29,6 +46,8 @@ def index(request):
 
     return render( request, 'beekeeping_app/index.html', {'Apiaries': all_apiarys})
 
+@login_required(login_url='login')
+@allowed_users(allowed_roles=['beekeeper_role'])
 def keeperView(request):
     all_keepers = Keeper.objects.all()
     #pass all keepers to the html file 
@@ -46,31 +65,45 @@ def keeperDetail(request, keeper):
 def hiveDetail(request, hive):
     hive_instance = Hive.objects.get(pk=hive)
     #for calendar:
-    events = Event.objects.filter(hive=hive)
-    return render(request, 'beekeeping_app/hive-detail.html', {'hive': hive, 'events': events})
+    #events = Event.objects.filter(hive=hive)
+    #return render(request, 'beekeeping_app/hive-detail.html', {'hive': hive, 'events': events})
+    return render(request, 'beekeeping_app/hive-detail.html', {'hive': hive_instance})
 
 
 #new should show an empty form - when submitted it adds that hive to the keeper's apiary
-def newHive(request, apiary):
-    apiary_instance = Apiary.objects.get(pk=apiary)
+@login_required(login_url='login')
+@allowed_users(allowed_roles=['beekeeper_role'])
+def newHive(request, apiary_id):
+    apiary_instance = Apiary.objects.get(pk=apiary_id)
 
+    #can do this because of our reverse reference
+    if apiary_instance.keeper.user != request.user:
+        raise PermissionDenied("You don't have permission to edit this hive.")
+    
     if request.method == "POST":
         form = HiveForm(request.POST)
         if form.is_valid():
             hive = form.save(commit=False)
             hive.apiary = apiary_instance  # Associate the hive with the apiary
             hive.save()
-            return redirect('hive-detail', hive.id)
+            return redirect('hive-detail', hive_id=hive.id)
+        
     else:
         form = HiveForm()
 
     return render(request, 'beekeeping_app/create_hive.html', {'apiary': apiary_instance, 'form': form})
 
+
 #update should show the existing hive with the fields filled and editable
+@login_required(login_url='login')
+@allowed_users(allowed_roles=['beekeeper_role'])
 def updateHive(request, apiary, hive):
     apiary_instance = Apiary.objects.get(id=apiary)
     hive_instance = Hive.objects.get(id=hive)
 
+    if apiary_instance.keeper.user != request.user:
+        raise PermissionDenied("You don't have permission to edit this hive.")
+    
     if request.method == "POST":
         form = HiveForm(request.POST, instance=hive_instance)
         if form.is_valid():
@@ -84,9 +117,14 @@ def updateHive(request, apiary, hive):
 
 #delete should show a 'Are you sure you want to delete 'Hive Name'?' 
 #Cancel (brings back to apiary-detail) Submit (deletes the hive)
+@login_required(login_url='login')
+@allowed_users(allowed_roles=['beekeeper_role'])
 def deleteHive(request, apiary, hive):
+    apiary_instance = Apiary.objects.get(id=apiary)
     hive_instance = Hive.objects.get(id=hive)
-    
+    if apiary_instance.keeper.user != request.user:
+        raise PermissionDenied("You don't have permission to edit this hive.")
+
     if request.method == "POST":
         hive_instance.delete()
         #redirecting to the apiary again after deletion
@@ -96,9 +134,13 @@ def deleteHive(request, apiary, hive):
 
 
 #update should show the existing hive with the fields filled and editable
+@login_required(login_url='login')
+@allowed_users(allowed_roles=['beekeeper_role'])
 def updateApiary(request, keeper, apiary):
     apiary_instance = Apiary.objects.get(id=apiary)
-
+    if apiary_instance.keeper.user != request.user:
+        raise PermissionDenied("You don't have permission to edit this apiary.")
+    
     if request.method == "POST":
         form = ApiaryForm(request.POST, request.FILES, instance=apiary_instance)
         if form.is_valid():
@@ -125,28 +167,67 @@ def updateApiary(request, keeper, apiary):
 
 #calendar views:
 
-class CalendarView(generic.ListView):
-    model = Event
-    template_name = 'cal/calendar.html'
+# class CalendarView(generic.ListView):
+#     model = Event
+#     template_name = 'cal/calendar.html'
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
+#     def get_context_data(self, **kwargs):
+#         context = super().get_context_data(**kwargs)
 
-        # use today's date for the calendar
-        d = get_date(self.request.GET.get('day', None))
+#         # use today's date for the calendar
+#         d = get_date(self.request.GET.get('day', None))
 
-        # Instantiate our calendar class with today's year and date
-        cal = Calendar(d.year, d.month)
+#         # Instantiate our calendar class with today's year and date
+#         cal = Calendar(d.year, d.month)
 
-        # Call the formatmonth method, which returns our calendar as a table
-        html_cal = cal.formatmonth(withyear=True)
-        context['calendar'] = mark_safe(html_cal)
-        return context
+#         # Call the formatmonth method, which returns our calendar as a table
+#         html_cal = cal.formatmonth(withyear=True)
+#         context['calendar'] = mark_safe(html_cal)
+#         return context
 
-def get_date(req_day):
-    if req_day:
-        year, month = (int(x) for x in req_day.split('-'))
-        return datetime.date(year, month, day=1)
-    return datetime.today()
+# def get_date(req_day):
+#     if req_day:
+#         year, month = (int(x) for x in req_day.split('-'))
+#         return datetime.date(year, month, day=1)
+#     return datetime.today()
 
+#authentication for registering a user:
+def registerPage(request):
+    form = CreateUserForm()
+
+    if request.method == 'POST':
+        form = CreateUserForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            username = form.cleaned_data.get('username')
+            # Create a new group 'beekeeper_role' if it doesn't exist
+            group, created = Group.objects.get_or_create(name='beekeeper_role')
+            # Add the user to the 'beekeeper_role' group
+            group.user_set.add(user)
+            apiary = Apiary.objects.create()
+            keeper = Keeper.objects.create(user=user, apiary=apiary)
+            messages.success(request, 'Account was created for ' + username)
+            return redirect('login')
+    
+    context = {'form':form}
+    return render(request, 'registration/register.html', context)
+
+@login_required(login_url='login')
+@allowed_users(allowed_roles=['student'])
+def userPage(request):
+    keeper = request.user.keeper
+    form = BeekeeperForm(instance = keeper)
+    print('keeper', keeper)
+    apiary = keeper.apiary
+    print(apiary)
+    if request.method == 'POST':
+        form = BeekeeperForm(request.POST, request.FILES, instance=keeper)
+        if form.is_valid():
+            form.save()
+    context = {'apiaries':apiary, 'form':form}
+    return render(request, 'beekeeping_app/user.html', context)
+
+def logoutView(request):
+    logout(request)
+    return render(request, 'registration/logout.html')
